+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Sin Publicar
 - Added Multiple Row Component
 - Resized pill icon on Hybrid Carousel
+- Added Multiple Description Component
+- Added Multiple Description to Row
 - New image accesory for Hyrbrid Carousel Card
 
 # v1.18.0